requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.4
Pillow==6.2.1
pytesseract==0.3.0
Werkzeug==0.16.0
